fprint progress

+++ include/f_math.h
@@ -6,15 +6,15 @@
 /*   By: wseegers <wseegers.mauws@gmail.com>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/28 11:53:54 by wseegers          #+#    #+#             */
-/*   Updated: 2018/05/31 20:26:44 by wseegers         ###   ########.fr       */
+/*   Updated: 2018/06/07 03:29:00 by wseegers         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef F_MATH_H
 # define F_MATH_H
 
-int		f_abs(int nbr);
-long	f_min(long base, long test);
-long	f_max(long base, long test);
+long long	f_abs(long long nbr);
+long		f_min(long base, long test);
+long		f_max(long base, long test);
 
 #endif
